==> step2_tokenizing_scRNAseq_data.ipynb <==
PROJECT_NAME = "reverse-gene-finder"

import os
PROJECT_HOME = os.path.join("/content/drive/My Drive/Projects", PROJECT_NAME)

import sys
sys.path.append(PROJECT_HOME)

# Google Drive storage setup
from google.colab import drive
drive.mount('/content/drive', force_remount=True)

%pip install anndata scanpy loompy datasets > /dev/null 2> /dev/null

from libs.tokenizer import TranscriptomeTokenizer

h5ad_data_dir = os.path.join(PROJECT_HOME, "data", "h5ad")
token_data_dir = os.path.join(PROJECT_HOME, "data", "tokenized_data")

if not os.path.exists(token_data_dir):
    os.makedirs(token_data_dir)

tokenizer = TranscriptomeTokenizer(nproc=8, custom_attr_name_dict={"joinid": "joinid", "disease": "disease"})
tokenizer.tokenize_data(
    data_directory=h5ad_data_dir,
    output_directory=token_data_dir,
    output_prefix="rosmap",
    file_format="h5ad"
)
