==> step4b_fine_tuning_final_evaluation.ipynb <==
PROJECT_NAME = "reverse-gene-finder"

import os
os.environ['SHELL'] = '/bin/bash'
PROJECT_HOME = os.path.join("/content/drive/My Drive/Projects", PROJECT_NAME)

import sys
sys.path.append(PROJECT_HOME)

# Google Drive storage setup
from google.colab import drive
drive.mount('/content/drive', force_remount=True)

import json
import numpy as np

sensitivity = []
specificity = []
roc_auc = []
for cv_idx in range(5):
    with open(os.path.join(PROJECT_HOME, 'results', 'eval_cv_%d.json' % cv_idx), 'r') as f:
        eval_results = json.load(f)
        sensitivity.append(eval_results['sensitivity'])
        specificity.append(eval_results['specificity'])
        roc_auc.append(eval_results['roc_auc'])

print("ROC AUC: %.2f%% ± %.2f%%" % (100 * np.mean(roc_auc), 100 * np.std(roc_auc)))
print("Sensitivity: %.2f%% ± %.2f%%" % (100 * np.mean(sensitivity), 100 * np.std(sensitivity)))
print("Specificity: %.2f%% ± %.2f%%" % (100 * np.mean(specificity), 100 * np.std(specificity)))
